Crate massa_pos_exports

Definition and exports of the PoS types and errors.

Define also the Selector worker that compute in background the draws for the future cycles

Modules

config 🔒

controller_traits 🔒

This module exports generic traits representing interfaces for interacting with the PoS selector worker.

cycle_info 🔒

deferred_credits 🔒

error 🔒

pos_changes 🔒

pos_final_state 🔒

settings 🔒

Macros

complete_key

Complete key formatting macro

deferred_credits_key

Deferred credits key formatting macro

final_state_hash_snapshot_key

Final state hash snapshot key formatting macro

prod_stats_fail_key

Production stats fail key formatting macro

prod_stats_prefix

Production stats prefix macro

prod_stats_success_key

Production stats success key formatting macro

rng_seed_key

Rng seed key formatting macro

roll_count_key

Roll count key formatting macro

roll_count_prefix

Roll count key prefix macro

Structs

CreditsDeserializer

Deserializer for a single credit

CreditsSerializer

Serializer for Credits

CycleHistoryDeserializer

Deserializer for cycle history, useful when restarting from a snapshot

CycleHistorySerializer

Serializer for cycle history

CycleInfo

State of a cycle for all threads

CycleInfoDeserializer

Deserializer for CycleInfo

CycleInfoSerializer

Serializer for CycleInfo

DeferredCredits

Structure containing all the PoS deferred credits information

DeferredCreditsDeserializer

Deserializer for DeferredCredits

DeferredCreditsSerializer

Serializer for DeferredCredits

PoSChanges

Recap of all PoS changes

PoSChangesDeserializer

PoSChanges Deserializer

PoSChangesSerializer

PoSChanges Serializer

PoSConfig

proof-of-stake final state configuration

PoSFinalState

Final state of PoS

ProductionStats

Block production statistics

ProductionStatsDeserializer

Deserializer for ProductionStats

ProductionStatsSerializer

Serializer for ProductionStats

RollsDeserializer

Deserializer for rolls

Selection

Selections of endorsements and producer

SelectorConfig

Configuration of selector thread

Enums

PosError

Proof-of-Stake error

Traits

SelectorController

interface that communicates with the selector worker thread

SelectorManager

Selector manager used to stop the selector thread

Type Aliases

PosResult

Proof-of-Stake result