Struct massa_pos_exports::pos_final_state::PoSFinalState

pub struct PoSFinalState {
    pub config: PoSConfig,
    pub db: ShareableMassaDBController,
    pub cycle_history_cache: VecDeque<(u64, bool)>,
    pub rng_seed_cache: Option<(u64, BitVec<u8>)>,
    pub selector: Box<dyn SelectorController>,
    pub initial_rolls: BTreeMap<Address, u64>,
    pub initial_seeds: Vec<Hash>,
    pub deferred_credits_serializer: DeferredCreditsSerializer,
    pub deferred_credits_deserializer: DeferredCreditsDeserializer,
    pub cycle_info_serializer: CycleHistorySerializer,
    pub cycle_info_deserializer: CycleHistoryDeserializer,
}

Final state of PoS

Fields
config: PoSConfig

proof-of-stake configuration

db: ShareableMassaDBController

Access to the RocksDB database

cycle_history_cache: VecDeque<(u64, bool)>

contiguous cycle history, back = newest

rng_seed_cache: Option<(u64, BitVec<u8>)>

rng_seed cache to get rng_seed for the current cycle

selector: Box<dyn SelectorController>

selector controller

initial_rolls: BTreeMap<Address, u64>

initial rolls, used for negative cycle look back

initial_seeds: Vec<Hash>

initial seeds, used for negative cycle look back (cycles -2, -1 in that order)

deferred_credits_serializer: DeferredCreditsSerializer

deferred credits serializer

deferred_credits_deserializer: DeferredCreditsDeserializer

deferred credits deserializer

cycle_info_serializer: CycleHistorySerializer

cycle info serializer

cycle_info_deserializer: CycleHistoryDeserializer

cycle info deserializer

Implementations

impl PoSFinalState

pub fn new( config: PoSConfig, initial_seed_string: &str, initial_rolls_path: &PathBuf, selector: Box<dyn SelectorController>, db: ShareableMassaDBController, ) -> Result<Self, PosError>

create a new PoSFinalState

pub fn load_initial_deferred_credits( &mut self, batch: &mut DBBatch, ) -> Result<(), PosError>

Try load initial deferred credits from file

pub fn recompute_pos_state_caches(&mut self)

After bootstrap or load from disk, recompute the caches

pub fn reset(&mut self)

Reset the state of the PoS final state

USED ONLY FOR BOOTSTRAP

pub fn create_initial_cycle(&mut self, batch: &mut DBBatch)

Create the initial cycle based off the initial rolls.

This should be called only if bootstrap did not happen.

pub fn create_new_cycle_from_last( &mut self, last_cycle_info: &CycleInfo, first_slot: Slot, last_slot: Slot, batch: &mut DBBatch, ) -> Result<(), PosError>

Create the a cycle based off of another cycle_info.

Used for downtime interpolation, when restarting from a snapshot.

pub fn delete_cycle_info(&mut self, cycle: u64, batch: &mut DBBatch)

Deletes a given cycle from RocksDB

pub fn compute_initial_draws(&mut self) -> PosResult<()>

Sends the current draw inputs (initial or bootstrapped) to the selector. Waits for the initial draws to be performed.

pub fn apply_changes_to_batch( &mut self, changes: PoSChanges, slot: Slot, feed_selector: bool, batch: &mut DBBatch, ) -> PosResult<()>

Technical specification of apply_changes_to_batch:

set self.last_final_slot = C if cycle C is absent from self.cycle_history_cache: push a new empty CycleInfo on disk and reflect in self.cycle_history_cache and set its cycle = C pop_front from cycle_history_cache until front() represents cycle C-4 or later (not C-3 because we might need older endorsement draws on the limit between 2 cycles) delete the removed cycles from disk for the cycle C entry in the db: extend seed_bits with changes.seed_bits extend roll_counts with changes.roll_changes delete all entries from roll_counts for which the roll count is zero add each element of changes.production_stats to the cycle’s production_stats for each changes.deferred_credits targeting cycle Ct: overwrite self.deferred_credits entries of cycle Ct in cycle_history with the ones from change remove entries for which Amount = 0 if slot S was the last of cycle C: set complete=true for cycle C in the history compute the seed hash and notifies the PoSDrawer for cycle C+3

pub fn feed_selector(&self, draw_cycle: u64) -> PosResult<()>

Feeds the selector targeting a given draw cycle

pub fn feed_cycle_state_hash( &self, cycle: u64, final_state_hash: HashXof<HASH_XOF_SIZE_BYTES>, )

Feeds the selector targeting a given draw cycle

impl PoSFinalState

pub fn get_rolls_for(&self, addr: &Address) -> u64

Retrieves the amount of rolls a given address has at the latest cycle

pub fn get_address_active_rolls( &self, addr: &Address, cycle: u64, ) -> Option<u64>

Retrieves the amount of rolls a given address has at a given cycle - 3 if cycle - 3 does not exist, values from initial rolls are returned

pub fn get_all_active_rolls(&self, cycle: u64) -> BTreeMap<Address, u64>

Gets all active rolls for a given cycle - 3, use self.initial_rolls if cycle - 3 does not exist

pub fn get_deferred_credits_range<R>( &self, range: R, addr_filter: Option<&Address>, ) -> DeferredCredits
where R: RangeBounds<Slot>,

Retrieves every deferred credit in a slot range Warning: this can be quite slow if there is tons of Deferred credits (e.g. just after genesis) to fetch so there is an option to filter by an address to improve the speed (cf json rpc api get_addresses)

pub fn get_cycle_index(&self, cycle: u64) -> Option<usize>

Gets the index of a cycle in history

pub fn get_all_roll_counts(&self, cycle: u64) -> BTreeMap<Address, u64>

Get all the roll counts for a given cycle

pub fn get_all_production_stats( &self, cycle: u64, ) -> Option<PreHashMap<Address, ProductionStats>>

Retrieves the productions statistics for all addresses on a given cycle

fn get_all_production_stats_private( &self, cycle: u64, ) -> PreHashMap<Address, ProductionStats>

Retrieves the productions statistics for all addresses on a given cycle

fn get_cycle_history_rng_seed(&self, cycle: u64) -> Option<BitVec<u8>>

Getter for the rng_seed of a given cycle, prioritizing the cache and querying the database as fallback.

fn get_cycle_history_final_state_hash_snapshot( &self, cycle: u64, ) -> Option<HashXof<HASH_XOF_SIZE_BYTES>>

Getter for the final_state_hash_snapshot of a given cycle.

Panics if the cycle is not in the history.

fn get_cycle_history_cycles(&self) -> Vec<(u64, bool)>

Used to recompute the cycle cache from the disk.

pub fn get_cycle_info(&self, cycle: u64) -> Option<CycleInfo>

Queries a given cycle info in the database

pub fn get_address_credits_for_slot( &self, addr: &Address, slot: &Slot, ) -> Option<Amount>

Gets the deferred credits for a given address that will be credited at a given slot

pub fn get_production_stats_for_address( &self, cycle: u64, address: &Address, ) -> Option<ProductionStats>

Gets the production stats for a given address

pub fn is_cycle_complete(&self, cycle: u64) -> Option<bool>

Check if a cycle is complete (all slots finalized)

impl PoSFinalState

pub fn put_new_cycle_info( &mut self, cycle_info: &CycleInfo, batch: &mut DBBatch, )

Helper function to put a new CycleInfo to RocksDB, and update the cycle_history cache

fn put_cycle_history_complete( &mut self, cycle: u64, value: bool, batch: &mut DBBatch, )

Helper function to put a the complete flag for a given cycle

fn put_cycle_history_final_state_hash_snapshot( &self, cycle: u64, value: Option<HashXof<HASH_XOF_SIZE_BYTES>>, batch: &mut DBBatch, )

Helper function to put a the final_state_hash_snapshot for a given cycle

fn put_cycle_history_rng_seed( &mut self, cycle: u64, value: BitVec<u8>, batch: &mut DBBatch, )

Helper function to put a the rng_seed for a given cycle

fn put_cycle_history_address_entry( &self, cycle: u64, address: &Address, roll_count: Option<&u64>, production_stats: Option<&ProductionStats>, batch: &mut DBBatch, )

Internal function to put an entry for a given address in the cycle history

pub fn put_deferred_credits_entry( &self, slot: &Slot, address: &Address, amount: &Amount, batch: &mut DBBatch, )

Internal function to put an entry

impl PoSFinalState

Helpers for key and value management

fn cycle_history_cycle_prefix(&self, cycle: u64) -> Vec<u8>

Helper function to construct the key prefix associated with a given cycle

pub fn is_cycle_history_key_value_valid( &self, serialized_key: &[u8], serialized_value: &[u8], ) -> bool

Deserializes the key and value, useful after bootstrap

pub fn is_deferred_credits_key_value_valid( &self, serialized_key: &[u8], serialized_value: &[u8], ) -> bool

Deserializes the key and value, useful after bootstrap