Struct massa_pos_exports::DeferredCredits

pub struct DeferredCredits {
    pub credits: BTreeMap<Slot, PreHashMap<Address, Amount>>,
    hash_tracker: Option<DeferredCreditsHashTracker>,
}

Structure containing all the PoS deferred credits information

Fields
credits: BTreeMap<Slot, PreHashMap<Address, Amount>>

Deferred credits

hash_tracker: Option<DeferredCreditsHashTracker>

Hash tracker, optional. Indeed, computing the hash is expensive, so we only compute it when finalizing a slot.

Implementations

impl DeferredCredits

pub fn is_empty(&self) -> bool

Check if the credits list is empty

pub fn new_with_hash() -> Self

Create a new DeferredCredits with hash tracking

pub fn new_without_hash() -> Self

Create a new DeferredCredits without hash tracking

pub fn get_hash(&self) -> Option<&Hash>

Get hash from tracker, if any

pub fn for_each<F>(&mut self, f: F)where F: FnMut(&Slot, &Address, &mut Amount),

Apply a function to each element

pub fn enable_hash_tracker_and_compute_hash(&mut self) -> &Hash

Enables the hash tracker (and compute the hash if absent)

pub fn disable_hash_tracker(&mut self)

Disable the hash tracker, loses hash

pub fn get_slot_range<R>(&self, range: R, with_hash: bool) -> DeferredCreditswhere R: RangeBounds<Slot>,

Get all deferred credits within a slot range. If with_hash == true then the resulting DeferredCredits contains the hash of the included data. Note that computing the hash is heavy and should be done only at finalization.

pub fn extend(&mut self, other: Self)

Extends the current DeferredCredits with another and replace the amounts for existing addresses

pub fn remove_zeros(&mut self)

Remove credits set to zero, use only on finality

pub fn get_address_credits_for_slot( &self, addr: &Address, slot: &Slot ) -> Option<Amount>

Gets the deferred credits for a given address that will be credited at a given slot

pub fn insert( &mut self, slot: Slot, address: Address, amount: Amount ) -> Option<Amount>

Insert an element