Struct massa_pos_exports::DeferredCredits

pub struct DeferredCredits {
    pub credits: BTreeMap<Slot, PreHashMap<Address, Amount>>,
    pub hash: Hash,
}

Structure containing all the PoS deferred credits information

Fields
credits: BTreeMap<Slot, PreHashMap<Address, Amount>>

Deferred credits

hash: Hash

Hash of the current deferred credits state

Implementations

impl DeferredCredits

pub fn nested_extend(&mut self, other: Self)

Extends the current DeferredCredits with another and replace the amounts for existing addresses

pub fn final_nested_extend(&mut self, other: Self)

Extends the current DeferredCredits with another, replace the amounts for existing addresses and compute the object hash, use only on finality

pub fn remove_zeros(&mut self)

Remove credits set to zero, use only on finality

pub fn get_address_deferred_credit_for_slot( &self, addr: &Address, slot: &Slot ) -> Option<Amount>

Gets the deferred credits for a given address that will be credited at a given slot

pub fn insert(&mut self, addr: Address, slot: Slot, amount: Amount)

Insert/overwrite a deferred credit