Struct massa_pos_exports::PoSFinalState

pub struct PoSFinalState {
    pub config: PoSConfig,
    pub cycle_history: VecDeque<CycleInfo>,
    pub deferred_credits: DeferredCredits,
    pub selector: Box<dyn SelectorController>,
    pub initial_rolls: BTreeMap<Address, u64>,
    pub initial_seeds: Vec<Hash>,
    pub initial_ledger_hash: Hash,
}

Final state of PoS

Fields
config: PoSConfig

proof-of-stake configuration

cycle_history: VecDeque<CycleInfo>

contiguous cycle history, back = newest

deferred_credits: DeferredCredits

coins to be credited at the end of the slot

selector: Box<dyn SelectorController>

selector controller

initial_rolls: BTreeMap<Address, u64>

initial rolls, used for negative cycle look back

initial_seeds: Vec<Hash>

initial seeds, used for negative cycle look back (cycles -2, -1 in that order)

initial_ledger_hash: Hash

initial state hash

Implementations

impl PoSFinalState

pub fn new( config: PoSConfig, initial_seed_string: &str, initial_rolls_path: &PathBuf, selector: Box<dyn SelectorController>, initial_ledger_hash: Hash ) -> Result<Self, PosError>

create a new PoSFinalState

pub fn from_snapshot( config: PoSConfig, cycle_history: VecDeque<CycleInfo>, deferred_credits: DeferredCredits, initial_seed_string: &str, initial_rolls_path: &PathBuf, selector: Box<dyn SelectorController>, initial_ledger_hash: Hash ) -> Result<Self, PosError>

create a PoSFinalState from an existing snapshot

pub fn reset(&mut self)

Reset the state of the PoS final state

USED ONLY FOR BOOTSTRAP

pub fn create_initial_cycle(&mut self)

Create the initial cycle based off the initial rolls.

This should be called only if bootstrap did not happen.

pub fn compute_initial_draws(&mut self) -> PosResult<()>

Sends the current draw inputs (initial or bootstrapped) to the selector. Waits for the initial draws to be performed.

pub fn apply_changes( &mut self, changes: PoSChanges, slot: Slot, feed_selector: bool ) -> PosResult<()>

Technical specification of apply_changes:

set self.last_final_slot = C if cycle C is absent from self.cycle_history: push a new empty CycleInfo at the back of self.cycle_history and set its cycle = C pop_front from cycle_history until front() represents cycle C-4 or later (not C-3 because we might need older endorsement draws on the limit between 2 cycles) for the cycle C entry of cycle_history: extend seed_bits with changes.seed_bits extend roll_counts with changes.roll_changes delete all entries from roll_counts for which the roll count is zero add each element of changes.production_stats to the cycle’s production_stats for each changes.deferred_credits targeting cycle Ct: overwrite self.deferred_credits entries of cycle Ct in cycle_history with the ones from change remove entries for which Amount = 0 if slot S was the last of cycle C: set complete=true for cycle C in the history compute the seed hash and notifies the PoSDrawer for cycle C+3

pub fn feed_cycle_state_hash(&mut self, cycle: u64, final_state_hash: Hash)

Feeds the selector targeting a given draw cycle

pub fn get_rolls_for(&self, addr: &Address) -> u64

Retrieves the amount of rolls a given address has at the latest cycle

pub fn get_address_active_rolls(&self, addr: &Address, cycle: u64) -> Option<u64>

Retrieves the amount of rolls a given address has at a given cycle

pub fn get_deferred_credits_at(&self, slot: &Slot) -> PreHashMap<Address, Amount>

Retrieves every deferred credit of the given slot

pub fn get_all_production_stats( &self, cycle: u64 ) -> Option<&PreHashMap<Address, ProductionStats>>

Retrieves the productions statistics for all addresses on a given cycle

pub fn get_cycle_index(&self, cycle: u64) -> Option<usize>

Gets the index of a cycle in history

pub fn get_cycle_history_part( &self, cursor: StreamingStep<u64> ) -> Result<(Option<CycleInfo>, StreamingStep<u64>), ModelsError>

Gets a cycle of the Proof of Stake cycle_history. Used only in the bootstrap process.

Arguments:

cursor: indicates the bootstrap state after the previous payload

Returns

The PoS cycle and the updated cursor

pub fn get_deferred_credits_part( &self, cursor: StreamingStep<Slot> ) -> (DeferredCredits, StreamingStep<Slot>)

Gets a part of the Proof of Stake deferred_credits. Used only in the bootstrap process.

Arguments:

cursor: indicates the bootstrap state after the previous payload

Returns

The PoS deferred_credits part and the updated cursor

pub fn set_cycle_history_part( &mut self, part: Option<CycleInfo> ) -> StreamingStep<u64>

Sets a part of the Proof of Stake cycle_history. Used only in the bootstrap process.

Arguments

part: a CycleInfo received from get_pos_state_part and used to update PoS final state

pub fn set_deferred_credits_part( &mut self, part: DeferredCredits ) -> StreamingStep<Slot>

Sets a part of the Proof of Stake deferred_credits. Used only in the bootstrap process.

Arguments

part: DeferredCredits from get_pos_state_part and used to update PoS final state