Trait massa_pos_exports::SelectorController

pub trait SelectorController: Send + Sync {
    // Required methods
    fn wait_for_draws(&self, cycle: u64) -> PosResult<u64>;
    fn feed_cycle(
        &self,
        cycle: u64,
        lookback_rolls: BTreeMap<Address, u64>,
        lookback_seed: Hash
    ) -> PosResult<()>;
    fn get_selection(&self, slot: Slot) -> PosResult<Selection>;
    fn get_producer(&self, slot: Slot) -> PosResult<Address>;
    fn get_available_selections_in_range<'a>(
        &self,
        slot_range: RangeInclusive<Slot>,
        restrict_to_addresses: Option<&'a PreHashSet<Address>>
    ) -> PosResult<BTreeMap<Slot, Selection>>;
    fn clone_box(&self) -> Box<dyn SelectorController>;
}

interface that communicates with the selector worker thread

Required Methods

fn wait_for_draws(&self, cycle: u64) -> PosResult<u64>

Waits for draws to reach at least a given cycle number. Returns the latest cycle number reached (can be higher than cycle). Errors can occur if the thread stopped.

fn feed_cycle( &self, cycle: u64, lookback_rolls: BTreeMap<Address, u64>, lookback_seed: Hash ) -> PosResult<()>

Feed cycle to the selector

Arguments

cycle: cycle number to be drawn

lookback_rolls: look back rolls used for the draw (cycle - 3)

lookback_seed: look back seed hash for the draw (cycle - 2)

fn get_selection(&self, slot: Slot) -> PosResult<Selection>

Get Selection computed for a slot

fn get_producer(&self, slot: Slot) -> PosResult<Address>

Get Address of the selected block producer for a given slot

fn get_available_selections_in_range<'a>( &self, slot_range: RangeInclusive<Slot>, restrict_to_addresses: Option<&'a PreHashSet<Address>> ) -> PosResult<BTreeMap<Slot, Selection>>

Get selections computed for a slot range (only returns available selections):

Arguments

slot_range: range of slots to get the selection for

restrict_to_addresses: optionally restrict only to slots involving a given address

fn clone_box(&self) -> Box<dyn SelectorController>

Returns a boxed clone of self. Useful to allow cloning Box<dyn SelectorController>.